refactor(deformation_tracking): split helpers into tracking_io/optim/remap and extract build_parser
Promote three reusable helper sets out of the previously 1400-line
deformation_tracking.py:

- fim.refactor.tracking_io     — TIFF / .npy / figure I/O (8 functions)
- fim.refactor.tracking_optim  — torch forward model + initial-shift estimate (5)
- fim.refactor.tracking_remap  — smoothing + Lagrangian reference-grid remap (4)

deformation_tracking.py now keeps only the CLI entry point, default TIFF
paths, and main() orchestration. The argparse setup is also promoted to a
module-level build_parser() to mirror main_VFM and enable defaults/choice
testing without launching the full pipeline.

Tests follow the source layout: per-helper-module test files
(test_tracking_io.py / test_tracking_optim.py / test_tracking_remap.py)
plus TestBuildParser, TestMainPipeline, and TestModuleMainGuard in the
driver test. 63 passed, 1 skipped, 100/100/100/99% coverage on the four
source files.

Verified bit-identical: every output .npy/.json/.png matches the
pre-refactor monolithic version on a synthetic 14x14x10 stack with the
same torch seed (only run_info.txt differs in absolute paths and
wall-clock timing fields, both expected by design).

Made-with: Cursor
